moving handlestudysubmit into studyform
