+++ web/src/components/studies/StudyForm.tsx
@@ -4,6 +4,7 @@ import Dialog from "../ui/Dialog";
 import { Input, Alert, AlertMessage, Label, HelperText, Textarea } from "../shared/exports";
 import styles from "./StudyForm.module.css";
 import { Controller, SubmitHandler, useForm, useWatch, useFieldArray } from "react-hook-form";
+import { postStudies, StudyCreateRequest, ValidationError } from "@/openapi";
 
 export type StudyFormData = {
   title: string;
@@ -38,19 +39,18 @@ export type StudyFormData = {
 type StudyProps = {
   username: string;
   setStudyFormOpen: (name: boolean) => void;
-  handleStudySubmit: (data: StudyFormData) => Promise<void>;
-  submitError: string | null;
-  isSubmitting: boolean;
-  setSubmitError: (error: string | null) => void;
+  fetchStudies?: () => void;
+  isUpdate?: boolean;
 };
 
 // data protection office id
 const UclDpoId = "Z6364106";
 
-const domainName = process.env.NEXT_PUBLIC_DOMAIN_NAME;
+// this should match the domain that is used for the entra ID users in the portal
+const domainName = process.env.NEXT_PUBLIC_DOMAIN_NAME || "@ucl.ac.uk";
 
 export default function StudyForm(StudyProps: StudyProps) {
-  const { username, setStudyFormOpen, handleStudySubmit, submitError, isSubmitting, setSubmitError } = StudyProps;
+  const { username, setStudyFormOpen, fetchStudies, isUpdate } = StudyProps;
   const {
     register,
     handleSubmit,
@@ -67,6 +67,8 @@ export default function StudyForm(StudyProps: StudyProps) {
       additionalStudyAdminUsernames: [],
     },
   });
+  const [isSubmitting, setIsSubmitting] = useState(false);
+  const [submitError, setSubmitError] = useState<string | null>(null);
 
   const [currentStep, setCurrentStep] = useState(1);
   const nextStep = () => setCurrentStep(currentStep + 1);
@@ -112,8 +114,78 @@ export default function StudyForm(StudyProps: StudyProps) {
     }
   }, [controllerValue, setValue]);
 
+  // move this to study form & have it determine if it's an update if an id is given for the study
+  const handleStudySubmitCreate = async (data: StudyFormData) => {
+    setIsSubmitting(true);
+    setSubmitError(null);
+
+    try {
+      // Convert form data to StudyCreateRequest API format
+      const studyData: StudyCreateRequest = {
+        title: data.title,
+        description: data.description ? data.description : undefined,
+        data_controller_organisation: data.dataControllerOrganisation.toLowerCase(),
+        additional_study_admin_usernames: data.additionalStudyAdminUsernames
+          .map((admin) => admin.value.trim())
+          .map((username) => `${username}${domainName}`),
+        involves_ucl_sponsorship: data.involvesUclSponsorship ? data.involvesUclSponsorship : undefined,
+        involves_cag: data.involvesCag ? data.involvesCag : undefined,
+        cag_reference: data.cagReference ? data.cagReference.toString() : undefined,
+        involves_ethics_approval: data.involvesEthicsApproval ? data.involvesEthicsApproval : undefined,
+        involves_hra_approval: data.involvesHraApproval ? data.involvesHraApproval : undefined,
+        iras_id: data.irasId ? data.irasId : undefined,
+        is_nhs_associated: data.isNhsAssociated ? data.isNhsAssociated : undefined,
+        involves_nhs_england: data.involvesNhsEngland ? data.involvesNhsEngland : undefined,
+        nhs_england_reference: data.nhsEnglandReference ? data.nhsEnglandReference.toString() : undefined,
+        involves_mnca: data.involvesMnca ? data.involvesMnca : undefined,
+        requires_dspt: data.requiresDspt ? data.requiresDspt : undefined,
+        requires_dbs: data.requiresDbs ? data.requiresDbs : undefined,
+        is_data_protection_office_registered: data.isDataProtectionOfficeRegistered
+          ? data.isDataProtectionOfficeRegistered
+          : undefined,
+        data_protection_number:
+          data.dataProtectionPrefix && data.dataProtectionDate && data.dataProtectionId
+            ? `${data.dataProtectionPrefix}/${data.dataProtectionDate}/${data.dataProtectionId}`
+            : undefined,
+        involves_third_party: data.involvesThirdParty ? data.involvesThirdParty : undefined,
+        involves_external_users: data.involvesExternalUsers ? data.involvesExternalUsers : undefined,
+        involves_participant_consent: data.involvesParticipantConsent ? data.involvesParticipantConsent : undefined,
+        involves_indirect_data_collection: data.involvesIndirectDataCollection
+          ? data.involvesIndirectDataCollection
+          : undefined,
+        involves_data_processing_outside_eea: data.involvesDataProcessingOutsideEea
+          ? data.involvesDataProcessingOutsideEea
+          : undefined,
+      };
+
+      const response = await postStudies({
+        body: studyData,
+      });
+
+      if (response.data && fetchStudies) {
+        setStudyFormOpen(false);
+        fetchStudies();
+        return;
+      }
+
+      if (response.error) {
+        const errorData = response.error as ValidationError;
+        if (errorData?.error_message) {
+          setSubmitError(errorData.error_message);
+          return;
+        }
+      }
+
+      setSubmitError("An unknown error occurred.");
+    } catch (error) {
+      console.error("Failed to create study:", error);
+      setSubmitError("Failed to create study. Please try again.");
+    } finally {
+      setIsSubmitting(false);
+    }
+  };
   const onSubmit: SubmitHandler<StudyFormData> = async (data) => {
-    await handleStudySubmit(data);
+    if (!isUpdate) await handleStudySubmitCreate(data);
   };
 
   const handleCloseForm = () => {
